functions/requirements.txt
Flask==2.0.1
google-cloud-storage==1.42.3
firebase-admin==5.0.3
google-cloud-firestore==2.1.0
werkzeug==2.0.3
azure-cognitiveservices-vision-computervision
langchain
langchain-anthropic
langchain-openai
langchain-community
PyMuPDF
spacy
transformers
sentence-transformers
scikit-learn
nltk
python-dotenv
numpy
pandas
google-resumable-media==1.3.0
pydantic
easyocr==1.7.1
opencv-python-headless==4.10.0.84
reportlab